requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pyyaml>=6.0.1
